+++ CHANGES.md
@@ -1,6 +1,23 @@
 Changelog
 =========
 
+0.3.0 (UNRELEASED)
+------------------
+
+### Added
+
+- Configuration field `keep_timeboxes` to configure the maximum number
+  of timeboxes to retain in state per user per channel.
+- Configuration field `keep_duration` to configure the maximum
+  duration for which timeboxes are retained in state.
+- Configuration field `max_print_channel` to configure the maximum
+  number of timeboxes to be listed in a channel in response to `list`
+  or `mine` commands.
+- Configuration field `max_print_private` to configure the maximum
+  number of timeboxes to be listed in private message in response to
+  `list` or `mine` commands.
+
+
 0.2.0 (2024-08-03)
 ------------------